uh oh commit

+++ cpp/include/hypercube/db/atom_cache.hpp
@@ -9,9 +9,10 @@
 namespace hypercube::db {
 
 // Cached atom information for fast lookup during ingestion
+// Coordinates are unsigned 32-bit values from YOUR Hilbert mapping
 struct AtomInfo {
     Blake3Hash hash;
-    int32_t coord_x, coord_y, coord_z, coord_m;
+    int64_t coord_x, coord_y, coord_z, coord_m;  // BIGINT from DB (unsigned 32-bit fits in signed 64-bit)
 };
 
 // Load atoms for specific codepoints only (fast - queries only needed atoms)
